Added calling login function
